Change integrate_rotor_system function to be a rotor method
