+++ ross/utils.py
@@ -4,7 +4,6 @@ import numpy as np
 import pandas as pd
 from numpy import linalg as la
 from plotly import graph_objects as go
-from scipy.linalg import eigh
 
 
 class DataNotFoundError(Exception):
@@ -690,7 +689,8 @@ def newmark(fun, t, y_size, **options):
     >>> yout = newmark(rotor_system, t, rotor.ndof)
     >>> dof = 13
     >>> yout[:, dof] # doctest: +ELLIPSIS
-    array([0.0000000e+00, 8.4914005e-09, 4.3429676e-08, ...
+    array([0.00000000e+00, 8.49140057e-09, 4.34296767e-08, ...,
+           1.16148468e-05, 1.16492353e-05, 1.16859622e-05])
     """
 
     gamma = options.get("gamma", 0.5)
@@ -705,7 +705,7 @@ def newmark(fun, t, y_size, **options):
     ydot0 = np.zeros(ny)
     y2dot0 = np.zeros(ny)
 
-    yout = np.full((n_steps, ny), 1e-38, dtype=np.float32)
+    yout = np.full((n_steps, ny), 1e-38, dtype=t.dtype)
     yout[0, :] = y0
 
     for step in range(1, n_steps):
@@ -746,70 +746,6 @@ def newmark(fun, t, y_size, **options):
         yout[step, :] = y
 
     return yout
-
-
-def apply_pseudo_modal(rotor, speed, num_modes):
-    """Pseudo-modal method.
-
-    This method can be used to apply modal transformation to reduce model
-    of the rotor system.
-
-    Parameters
-    ----------
-    rotor: rs.Rotor
-        The rotor object.
-    speed : float
-        Rotor speed.
-    num_modes : int
-        The number of eigenvectors to consider in the modal transformation
-        with model reduction.
-
-    Returns
-    -------
-    matrix_to_modal : callable
-        Function to transform a square matrix from physical to modal space.
-    vector_to_modal : callable
-        Function to transform a vector from physical to modal space.
-    vector_from_modal : callable
-        Function to transform a vector from modal to physical space.
-
-    Examples
-    --------
-    >>> import ross as rs
-    >>> rotor = rs.rotor_example()
-    >>> size = 10000
-    >>> node = 3
-    >>> speed = 500.0
-    >>> t = np.linspace(0, 10, size)
-    >>> F = np.zeros((size, rotor.ndof))
-    >>> F[:, rotor.number_dof * node] = 10 * np.cos(2 * t)
-    >>> F[:, rotor.number_dof * node + 1] = 10 * np.sin(2 * t)
-    >>> get_array = apply_pseudo_modal(rotor, speed, num_modes=12)
-    >>> F_modal = get_array[1](F.T).T
-    >>> round(la.norm(F_modal), 5)
-    226.92798
-    """
-
-    M = rotor.M(speed)
-    K_aux = rotor.K(speed)
-
-    # Remove cross-coupled coefficients of bearing stiffness matrix
-    rmv_cross_coeffs = [[0, 1], [1, 0]]
-    if rotor.number_dof == 6:
-        rmv_cross_coeffs = [[0, 1, 0], [1, 0, 0], [0, 0, 0]]
-
-    for elm in rotor.bearing_elements:
-        dofs = list(elm.dof_global_index.values())
-        K_aux[np.ix_(dofs, dofs)] -= elm.K(speed) * rmv_cross_coeffs
-
-    _, modal_matrix = eigh(K_aux, M)
-    modal_matrix = modal_matrix[:, :num_modes]
-
-    matrix_to_modal = lambda array: (modal_matrix.T @ array) @ modal_matrix
-    vector_to_modal = lambda array: modal_matrix.T @ array
-    vector_from_modal = lambda array: modal_matrix @ array
-
-    return matrix_to_modal, vector_to_modal, vector_from_modal
 
 
 def assemble_C_K_matrices(elements, C0, K0, speed=None):
@@ -873,161 +809,6 @@ def assemble_C_K_matrices(elements, C0, K0, speed=None):
             K0[np.ix_(dofs, dofs)] += elm.K()
 
     return C0, K0
-
-
-def integrate_rotor_system(rotor, speed, F, t, **kwargs):
-    """Time integration for a rotor system.
-
-    This method returns the time response for a rotor given a force, time and
-    speed based on time integration with the Newmark method.
-
-    Parameters
-    ----------
-    speed : float or array_like
-        Rotor speed.
-    F : ndarray
-        Force array (needs to have the same length as time array).
-    t : ndarray
-        Time array.
-    **kwargs : optional
-        Additional keyword arguments can be passed to define the parameters
-        of the Newmark method if it is used (e.g. gamma, beta, tol, ...).
-        See `newmark` for more details. Other optional arguments are listed
-        below.
-    num_modes : int, optional
-        If `num_modes` is passed as argument, the pseudo-modal method is applied reducing
-        the model to the chosen number of modes.
-    add_to_RHS : callable, optional
-        An optional function that computes and returns an additional array to be added to
-        the right-hand side of the equation of motion. This function should take arguments
-        corresponding to the current state of the rotor system, including the time step
-        number, displacements, and velocities. It should return an array of the same length
-        as the degrees of freedom of the rotor system (`rotor.ndof`). This function allows
-        for the incorporation of supplementary terms or external effects in the rotor system
-        dynamics beyond the specified force input during the time integration process.
-
-    Returns
-    -------
-    t : ndarray
-        Time values for the output.
-    yout : ndarray
-        System response.
-
-    Examples
-    --------
-    >>> import ross as rs
-    >>> rotor = rs.rotor_example()
-    >>> size = 10000
-    >>> node = 3
-    >>> speed = 500.0
-    >>> accel = 0.0
-    >>> t = np.linspace(0, 10, size)
-    >>> F = np.zeros((size, rotor.ndof))
-    >>> F[:, rotor.number_dof * node] = 10 * np.cos(2 * t)
-    >>> F[:, rotor.number_dof * node + 1] = 10 * np.sin(2 * t)
-    >>> t, yout = integrate_rotor_system(rotor, speed, F, t)
-    Running direct method
-    >>> dof = 13
-    >>> yout[:, dof] # doctest: +ELLIPSIS
-    array([0.0000000e+00, 8.4914005e-09, 4.3429676e-08, ...
-    """
-
-    try:
-        speed_is_array = len(set(speed)) > 1
-        speed_ref = np.mean(speed)
-
-    except:
-        speed_is_array = False
-        speed_ref = speed
-
-    num_modes = kwargs.get("num_modes")
-
-    if num_modes and num_modes > 0:
-        size = num_modes
-        print("Running pseudo-modal method, number of modes =", size)
-        get_array = apply_pseudo_modal(rotor, speed_ref, size)
-
-    else:
-        size = rotor.ndof
-        print("Running direct method")
-        return_array = lambda array: array
-        get_array = [return_array for j in range(3)]
-
-    add_to_RHS = kwargs.get("add_to_RHS")
-
-    if add_to_RHS is None:
-        ext_force = lambda i, u, v: 0
-
-    else:
-        ext_force = lambda i, u, v: get_array[1](
-            add_to_RHS(i, get_array[2](u), get_array[2](v))
-        )
-
-    M = get_array[0](rotor.M())
-    C2 = get_array[0](rotor.G())
-    K2 = get_array[0](rotor.Ksdt())
-    F = get_array[1](F.T).T
-
-    if speed_is_array:
-        accel = np.gradient(speed, t)
-
-        freq_is_none = False
-        for elm in rotor.bearing_elements:
-            freq_is_none = (elm.frequency is None) or freq_is_none
-
-        if freq_is_none:
-            C1 = get_array[0](rotor.C(speed_ref))
-            K1 = get_array[0](rotor.K(speed_ref))
-
-            rotor_system = lambda step, disp_resp, velc_resp: (
-                M,
-                C1 + C2 * speed[step],
-                K1 + K2 * accel[step],
-                F[step, :] + ext_force(step, disp_resp, velc_resp),
-            )
-
-        else:
-            elements_without_bearing = [
-                *rotor.shaft_elements,
-                *rotor.disk_elements,
-                *rotor.point_mass_elements,
-            ]
-
-            C0, K0 = assemble_C_K_matrices(
-                elements_without_bearing,
-                np.zeros((rotor.ndof, rotor.ndof)),
-                np.zeros((rotor.ndof, rotor.ndof)),
-            )
-
-            def rotor_system(step, disp_resp, velc_resp):
-                C, K = assemble_C_K_matrices(
-                    rotor.bearing_elements, np.copy(C0), np.copy(K0), speed[step]
-                )
-
-                C1 = get_array[0](C)
-                K1 = get_array[0](K)
-
-                return (
-                    M,
-                    C1 + C2 * speed[step],
-                    K1 + K2 * accel[step],
-                    F[step, :] + ext_force(step, disp_resp, velc_resp),
-                )
-
-    else:
-        C1 = get_array[0](rotor.C(speed_ref))
-        K1 = get_array[0](rotor.K(speed_ref))
-
-        rotor_system = lambda step, disp_resp, velc_resp: (
-            M,
-            C1 + C2 * speed,
-            K1,
-            F[step, :] + ext_force(step, disp_resp, velc_resp),
-        )
-
-    response = newmark(rotor_system, t, size, **kwargs)
-    yout = get_array[2](response.T).T
-    return t, yout
 
 
 def remove_axial_torsional_dofs(matrix_6dof):
